Move message templates into templates dir
